BGA-10 Move deserialization to transaction builder
